photoshare/backend/requirements.txt
django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.6
pillow==10.0.0
transformers==4.30.2
torch==2.0.1
django-environ==0.10.0